Reformatted code. Put the animation canvas into a separate file
