+++ client/src/components/BulletSection.tsx
@@ -1,6 +1,7 @@
 import React from "react";
 //@ts-ignore
 import anime from "animejs/lib/anime.es";
+import AnimationCanvas from "./AnimationCanvas";
 
 // a sample data for chosenEmoji
 const testEmojiData = {
@@ -19,7 +20,7 @@ class BulletSection extends React.Component<
   streakId: number = 0;
   emojiAnimations: any = {};
   emojiDiv: React.RefObject<HTMLDivElement> = React.createRef();
-  emojiCanvas: React.RefObject<HTMLCanvasElement> = React.createRef();
+  animeCanvas: React.RefObject<AnimationCanvas> = React.createRef();
   clickZone: React.RefObject<HTMLDivElement> = React.createRef();
   streakDisplay: React.RefObject<HTMLCanvasElement> = React.createRef();
   streakCount: React.RefObject<HTMLSpanElement> = React.createRef();
@@ -38,8 +39,6 @@ class BulletSection extends React.Component<
     };
   }
   componentDidMount() {
-    this.initializeCanvas();
-    this.initializeListener();
     this.initializeStreak();
   }
 
@@ -67,63 +66,12 @@ class BulletSection extends React.Component<
     }
   }
 
-  // will be used for future usage, detects clicks on the canvas
-  initializeListener() {
-    /*
-    const updateCoords = (e: any) => {
-      if (this.emojiCanvas.current) {
-        const rect = this.emojiCanvas.current.getBoundingClientRect();
-        const pointerX = (e.clientX || e.touches[0].clientX) - rect.left;
-        const pointerY = (e.clientY || e.touches[0].clientY) - rect.top;
-      }
-    };
-    */
-
-    if (this.emojiCanvas.current) {
-      this.emojiCanvas.current.addEventListener(
-        this.tap,
-        this.resetStreak,
-        false
-      );
-    }
-  }
-
   resetStreak = () => {
     this.setState({
       streakPoints: 0,
     });
-    console.log("reset streak");
     this.animateStreak();
   };
-
-  // The dimension of canvas matches the parent div element
-  initializeCanvas() {
-    const initializeDimension = () => {
-      if (this.emojiCanvas.current && this.emojiDiv.current) {
-        const canvasEl = this.emojiCanvas.current;
-        canvasEl.width = this.emojiDiv.current.clientWidth;
-        canvasEl.height = this.emojiDiv.current.clientHeight;
-        canvasEl.width = canvasEl.offsetWidth;
-        canvasEl.height = canvasEl.offsetHeight;
-      }
-    };
-
-    initializeDimension();
-    window.addEventListener("resize", initializeDimension);
-
-    if (this.emojiCanvas.current) {
-      const canvasEl = this.emojiCanvas.current;
-
-      // this clears the canvas screen on each update (avoid leaving the trace of animating objects)
-      const ctx = canvasEl.getContext("2d");
-      anime({
-        duration: Infinity,
-        update: function () {
-          ctx?.clearRect(0, 0, canvasEl.width, canvasEl.height);
-        },
-      });
-    }
-  }
 
   animateStreak() {
     if (this.streakDisplay.current) {
@@ -141,19 +89,11 @@ class BulletSection extends React.Component<
       countRef.style.color =
         "rgb(255," + Math.min(200, (this.state.streakPoints - 5) * 10) + ", 0)";
     }
-    var ml4 = {
-      opacityIn: [0, 1],
-      scaleIn: [0.2, 1],
-      scaleOut: 3,
-      durationIn: 800,
-      durationOut: 600,
-      delay: 500,
-    };
 
     anime.timeline().add({
       targets: ".streak-count-text",
-      scale: ml4.scaleIn,
-      duration: ml4.durationIn,
+      scale: [0.2, 1],
+      duration: 800,
     });
   }
 
@@ -225,68 +165,6 @@ class BulletSection extends React.Component<
     }, 50);
   }
 
-  createParticule(x: number, y: number) {
-    const colors = ["#FF1461", "#18FF92", "#5A87FF", "#FBF38C"];
-    if (this.emojiCanvas.current) {
-      const ctx = this.emojiCanvas.current.getContext("2d");
-      const p = {
-        x: x,
-        y: y,
-        color: colors[anime.random(0, colors.length - 1)],
-        radius: anime.random(16, 24),
-        endPos: this.setParticuleDirection(x, y),
-        draw: function () {
-          if (ctx) {
-            ctx.beginPath();
-            ctx.arc(p.x, p.y, p.radius, 0, 2 * Math.PI, true);
-            ctx.fillStyle = p.color;
-            ctx.fill();
-          }
-        },
-      };
-      return p;
-    }
-    return undefined;
-  }
-
-  setParticuleDirection(x: number, y: number) {
-    var angle = (anime.random(0, 360) * Math.PI) / 180;
-    var value = anime.random(25, 90);
-    var radius = [-1, 1][anime.random(0, 1)] * value;
-    return {
-      x: x + radius * Math.cos(angle),
-      y: y + radius * Math.sin(angle),
-    };
-  }
-
-  // create particles for the firework
-  renderParticule = (anim: any) => {
-    for (var i = 0; i < anim.animatables.length; i++) {
-      anim.animatables[i].target.draw();
-    }
-  };
-
-  // The animation for the firework
-  animateParticules = (x: number, y: number) => {
-    const particules = [];
-    for (var i = 0; i < 30; i++) {
-      particules.push(this.createParticule(x, y));
-    }
-    anime.timeline().add({
-      targets: particules,
-      x: function (p: any) {
-        return p.endPos.x;
-      },
-      y: function (p: any) {
-        return p.endPos.y;
-      },
-      radius: 0.1,
-      duration: anime.random(1200, 1800),
-      easing: "easeOutExpo",
-      update: this.renderParticule,
-    });
-  };
-
   addPoint() {
     const streakPoints = this.state.streakPoints + 1;
     const totalPoints =
@@ -307,7 +185,6 @@ class BulletSection extends React.Component<
         y >= 0 &&
         y <= rect.bottom - rect.top
       ) {
-        console.log("U got a point!");
         if (this.streakId !== parseInt(node.id) - 1) {
           this.resetStreak();
           this.streakId = parseInt(node.id);
@@ -325,7 +202,8 @@ class BulletSection extends React.Component<
     const x = parseFloat(
       transInfo.substring(transInfo.indexOf("(") + 1, transInfo.indexOf("px"))
     );
-    this.animateParticules(x, y);
+
+    this.animeCanvas.current?.animateParticules(x, y);
 
     this.withinClickZone(x, y, node);
     this.emojiAnimations[node.id].pause();
@@ -409,7 +287,7 @@ class BulletSection extends React.Component<
             // have a random time offset for each emoji (dont clutter together)
             setTimeout(() => {
               this.emojiToScreen(node);
-            }, Math.random() * 50);
+            }, Math.random() * 25);
           }
         });
       }
@@ -425,7 +303,10 @@ class BulletSection extends React.Component<
         </div>
         <div className="clickzone" ref={this.clickZone} />
         <div id="emojis" ref={this.emojiDiv}>
-          <canvas ref={this.emojiCanvas}></canvas>
+          <AnimationCanvas
+            ref={this.animeCanvas}
+            resetStreak={this.resetStreak}
+          />
           <canvas
             className="streak-container"
             ref={this.streakDisplay}
